requirements.txt
pyfiglet
pyautogui
os
Pillow